Safari-609.1.11.1/LayoutTests/fast/dom/timer-throttling-hidden-page-expected.txt

Tests that DOM timers gets throttled on hidden pages once they reach the max nesting level

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.isTimerThrottled(timerHandle) is false
PASS internals.isTimerThrottled(timerHandle) is false
PASS internals.isTimerThrottled(timerHandle) is false
PASS internals.isTimerThrottled(timerHandle) is false
PASS internals.isTimerThrottled(timerHandle) is false
PASS internals.isTimerThrottled(timerHandle) is true
PASS successfullyParsed is true

TEST COMPLETE
